requirements.txt
PySimpleGUI
requests
pillow
tls-client
typing-extensions